[vaapi][fuzz] Workaround one buffer collection
Current CodecImpl only supports one set of buffers allocated at a time.
This is a problem for the TiledSurfaceManager since reference frames
that are allocated prior to the buffer reallocation will be destroyed
and are unable to be used after. Added some temporary workaround logic
that detects if a non-keyframe requires a buffer reallocation and
ensures that the current implementation of surface manager supports it.

Bug: fxbug.dev/110529
Change-Id: I4816bd2608f71894a6c94c5c289b60e4fb80975e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753042
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>