[debugger] Parse and execute local var defs.
Parse a subset of C++ and Rust local variable definitions (see
variable_decl.cc for a description).

Create parse nodes for local variable declarations and local variable
uses. Emit bytecode for these constructs to execute them. Unlike
variables in the target, local variables are resolved at parse time and
converted to a location in the VM.

Add tests for declaring and using some simple local variables.

Adds an error and tests for the use of functional-style casts (which are
not supported).

Change-Id: I53fa563249b195f84e7b93f402782a11a69489ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748773
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>