[sys] Stricter threading model on OutgoingDirectory (part 1)
This CL is starts to tighten the threading restrictions on
OutgoingDirectory. With the eventual target being one must use the
outgoing directory from a dispatcher with mutual exclusion guarantee,
for example a single threaded async::Loop or a synchronized driver
dispatcher.

There are a couple of reasons for doing this:

- The underlying libsvc VFS implementation is thread-unsafe. It cannot
  be destroyed while the async dispatcher is running in the background.
- When someone calls RemoveProtocol etc., the expectation is that after
  the method returns, it's safe to destroy the server implementation
  passed to the initial AddProtocol. If the async dispatcher is running
  in the background, then we cannot guarantee this invariant -- the
  dispatcher might be servicing new connections to this very protocol.

While it is possible to destroy the OutgoingDirectory when the
dispatcher is paused (such as in-between two RunLoopUntilIdle), that
is very rare in production and subtle. This CL lets us start with a
more restrictive API.

This CL checks that the outgoing directory is created/destroyed/used
from the same thread/driver dispatcher. A follow-up one will check that
the dispatcher servicing the outgoing directory is the same one as the
one that created it. That one needs to sort out a few more clients that
were using the directory unsafely before it's ready.

Bug: 113997

Change-Id: I4bf3ec61315f95a053d4c77622d84d60bcec5aef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/752135
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>