[virtualization] Batch completions in the net device
Calling into the netstack is synchronous and can be slow, which is
why batched completions are supported. Conversely, waiting to batch
completions can add latency.

A compromise is to still send completions as quickly as possible, but
accrue completions in a queue for batching. For example, the C++ thread
schedules a "complete batches" task, which may be scheduled behind
multiple other tasks. While this waits to run, the Rust thread can
continue to append things to complete to the queue.

Note: This doesn't wire this up to the device, that'll be done in a
second CL to simplify the review and iperf testing.

Bug: 95485
Tested:
Unit tests

Change-Id: I1179a5a7cd92378c9936e3f7fc5d1661d7ffed57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/751042
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: David Hastings <dahastin@google.com>