[bt][hfp] Fix crash formatting codecs as strings
Previously, if the AG received a codec ID from a peer that was not one
specified in the HFP spec (either 0x1 or 0x2), it would panic attempting
to format the ID as a string.  This means that noncompliant peers can
crash HFP. We are seeing a few of these a day in the field.

This change adds support for unknown codecs to the string formatter for
code IDs and adds a test for this.

Test: fx test fuchsia-pkg://fuchsia.com/bt-hfp-audio-gateway-tests#meta/test-component.cm
Bug: fxb/113474
Change-Id: I276d00411f219b2947a9d2f0909c542b636204d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/752983
Commit-Queue: Matthew Kehrt <kehrt@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
(cherry picked from commit a0f251613dccea8556d7ce908a4663ebdd04bfc1)