[component_manager][events] Fix leak in event source
Spawn the event stream server task in the subscriber's task scope,
instead of the `task_scope` passed to `open` (which in this case is the
global one because event streams route to above-root).

This ensures that when the subscriber is destroyed, the task is
cancelled.

Bug: 110261
Change-Id: Ied39615fae0f512e2bf9aff537a789acf2650e4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/752984
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>