[component][fuzzing] Always pipe libFuzzer stdio
Within the Test Runner Framework, fuzzers that spam excessive amounts of
stdout and/or stderr have been observed to cause the output streams to
drop, leaving the forwarders in fuzz-manager to become stuck.

This CL addresses that problem by preventing the spam from reaching the
test runner in the first place. Except when retrying specific inputs,
stdout is suppressed. When retrying specific inputs, stdout and stderr
are both forwarded to provide as much diagnostics as possible.

Bug: 109100
Change-Id: Idfa78c49f99c726609bda8fd83dc8f9f5798372b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750504
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>