[basemgr] Fix basemgr lifecycle race.
basemgr was updated ~9 mos ago with a second implementation of process.lifecycle.Lifecycle. This implementation, which was bound to PA_LIFECYCLE handle, does not wait for BasemgrImpl's shutdown sequence to complete before stopping the main async loop.

The solution is to use the implementation built into BasemgrImpl, which calls an `on_shutdown` callback on completion of the shutdown sequence, which in turn stops the main loop.

Testing: added a unittest
Bug: 112090
Change-Id: I7f7fffcab32cbd6f70e093c9d6ec0d50c03f8b7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749287
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Gabe Schine <thatguy@google.com>