[image-format] Support bytes_per_row with Intel tiling
Intel Gen12 GPUs require that tiled images are multiples of 4x1 tiles
for use with color compression (even if the color compression buffer is
private to the driver and not exposed externally through sysmem). To
enforce that in sysmem, allow specifying bytes_per_row on Intel tiled
formats.

The "row size" is the number of bytes of a row of a linearized
view of a texture.

Bug: 109075
Change-Id: I4b17f72358ec1416a54e39af56b751f5323faa38
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748686
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>
API-Review: Dale Sather <dalesat@google.com>