[fhcp] Fail the touchpad test if erroneous input is detected
The test now assumes that the manual tester is infallible and that any
unexpected input (for example, a touch in the wrong quadrant) is the
result of a bug in the touchpad driver.

This allows for the simplification of some of the logic of the test.

Bug: b/255810966
Change-Id: Icd956031b9023ee64cc15ab37b386abff5f3868b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/752782
Reviewed-by: Harsha Priya N V <harshanv@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Ian Fisher <iafisher@google.com>