[kernel][x86][mmu] Make sure we put the CR4.PGE bit back after clearing it in a global tlb invalidate
As a result the first time we do a global invalidate on any given cpu,
the PGE bit was cleared and then never set again. As a result all global
pages were turned off for the rest of the system. Substantial
performance is being dropped on the floor as a result of this, since the
cpu now loses all kernel pages when it context switches between
processes.

Was wiped out in a previous refactoring of the code in February 2021.

Change-Id: I330d9e5b9ccf134f1bed08ee80621e834419460d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753522
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Carlos Pizano Uribe <cpu@google.com>