[pb] Move the MOS files outside the PB
The files intended for MOS (all_blobs, targets, images) must be uploaded
to GCS relative to the product bundle in the same way that it is
relative in the build directory. This is because these MOS files have
relative paths in them.

This CL addresses an issue where the images.json file was moved during
upload, causing the relative paths to be invalid. We now upload the MOS
files next to the product bundle just as they are uploaded to GCS.

Change-Id: Iadc58144ed2685b67b5d94d658f51a373d0b4ab1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753386
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>