[sdk] Fix depfile generation in cmc and use via _cmc_tool in the GN SDK
The cmc tool now emits a depfile for .cml files with no imports. This is
necessary, as otherwise the build system always determines that .cm
files must be rebuilt due to the absence of a depfile. This led to
unnecessary work during builds.

The cmc_compile template in the GN SDK properly emits a depfile, but the
_cmc_tool was failing to forward the `depfile` variable from its
invoker, so it wasn't actually being used in the build. This led to .cm
targets not being built when only included .cml files were modified.

Bug: 114310
Change-Id: I3d7354e550048c363eef13aa3dc81efdd9970555
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753627
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Greg Thompson <grt@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Sébastien Marchand <sebmarchand@google.com>