[login] Launch ermine/gazelle in a collection.
Instead of `ermine_shell` being a static child, create the application
shell as a dynamic child in a collection. Many of the changes in this
CL simply acknowledge the fact that the moniker has changed.

Besides changing the way the component gets created/started, this also
changes login.cm to proxy element.Manager and GraphicalPresenter
connections to the shell, since component manager can't do that for us
anymore.

Bug: fxbug.dev/100284

Change-Id: Id56b35115661d7eb051ef9b03d999e762453c765
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750667
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>