[ot-stack] Remove remnants from tree
This component is obsolete and we're not building it. If we keep it the
code will just rot, so we may as well remove it from the tree.

Removed in commit d5873301211ccdd21f90805928c23b212209af70.

Change-Id: Ib7cc26397368e6415852d3fc67b40912ef119849
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/754922
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>