[Workstation] Add initial UI for full settings app
This CL adds initial UI for full settings app. It features a selection menu on the left that lists the available settings and a detailed view on the right that shows the actively selected setting details/controls. Note that this isn't the final UI and various tweaks will be needed for proper scaling/usability.

Bug: 113485

Change-Id: I17991184e20dfe3620cf2fcbcc52f2884c6316ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753227
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Charles Whitten <cwhitten@google.com>