[bind] Define new USB_ASSOCIATION bind protocol
We ran into a situation where a host mode driver needs to distinguish
between binding on a raw interface, or an association of interfaces. It
would be preferable if this distinction could be made at bind time.

If binding on a device corresponding to an association, the
class/subclass/protocol are those of the IAD itself, while the interface
number corresponds to the first interface of the association. Otherwise,
if binding on a device corresponding to a raw interface, the
  class/subclass/protocol are those of the interface itself.

This change enables binding decisions to reflect interface vs.
association, as well as removes the need for a driver to dig through the
(possibly many) interfaces of the configuration to manually inspect
class/subclass/protocol discrepancies between the interface descriptor
and binding parameters.

Bug: 98137
Change-Id: I5c73e24536c74d9817031ee5b9899a2936ca6b05
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753342
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Sam Hansen <hansens@google.com>