[netstack3] Send close resp after socket is gone
Don't send the response to the final close request for a socket until
after the socket state is removed from Core. This prevents a race where
a client receives the close response and attempts to bind a socket to
the same address as the one closed, but the netstack hasn't yet removed
the state so the bind request fails.

Fixed: 112151
Change-Id: I800e0a04acae0b4828b2179abd858987b3fa903e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753985
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>