[netemul-runner-test] Fix flake
fxrev.dev/752984 changed the lifetime of event streams to bind them to
the lifetime of the subscriber.

netemul_test_runner grabs the test root's event stream from its
namespace. It expects the Destroyed event to be delivered over this
channel so that it knows when it can safely drop the network
environment. But when the test component is destroyed, component manager
closes the event stream, resulting in a race with PEER_CLOSED.

We can't revert fxrev.dev/752984 because it's needed to fix a leak in
component manager (which causes an even bigger flake). Instead, relax
the check to listen for PEER_CLOSED.

Multiply: netemul-runner-tests
Bug: 114292

Change-Id: I9c230eab9f4b43de6d3745a9e18db3c299c30baf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/755085
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>