[efi] Gracefully shutdown test applications
After UEFI test application execution, we want to be able to see the
output (and preferably not have it pushed off screen). When launched by
the shell, a simple exit does what we want: we return to the shell
without anything cleared or reset. Otherwise, not so: in QEMU, EDK II
will jump into the boot manager UI, clearing the screen - and on a small
sample of hardware, we boot loop. Accordingly, this change updates the
test exit logic to gracefully shut down after execution when not
launched from the shell.

Change-Id: I9320d501f2f3d12d9e7f34f113ba63c92f5f062b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/755668
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>