[boot-shim] Introduce DevicetreeBootShim
Add a simple shim with support for matchers that require
one |Match| iteration to reach completion, that is, the amount
of memory needed is statically known for the item.

Tests include one matcher of each type, plus multiple matchers.
These test are not intended to verify the appended item correctness,
but just the shim wiring.

Test: fx test boot-shim-unittests
Change-Id: I5d5f6576cde50bbea94dd33fe534b05903b18fb9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/751576
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>