[emu][mdns] Enable reading boot-time mdns configuration.
This change adds reading mdns service configuration from
the /boot data to allow overriding the default congfiguration.

This capability is used by the emulator when running with
user mode networking. In this mode, the emulator is running on
the local host network, and mapping ports on the guest (Fuchsia)
to unused ports on the host.

The overridden configuration includes text records that identify
the host IP address and port mappings being used so that tools
like `ffx` can discover emulator instances via mdns and connect
via the mapped ports.

Bug: 113136
Change-Id: I49960d0be7d39054d33bb10a437e1fb1d0558246
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/751011
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>