[zxdb] Remove dependency on "crashpad/client"
The Crashpad roller is blocked because its unable to build Crashpad's
client library for macOS. The library is only depended on by zxdb and
and the dependency is stale.

Removing the dependency _should_ unblock the roller because GN will no
longer need to build the client library.

Bug: 114328

Change-Id: I93280b4898dcf5309dad32c3a426a4f1ef4732f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/755930
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Alex Pankhurst <pankhurst@google.com>