[xhci] Additional logging when a TRB is dropped
Remove the assert for now to unblock workflows that previously sometimes
worked. We should add the assert back once we've unconvered all the root
causes for lost TRBs.

There's more data we could get here, but I think it would be best
gathered in inspect. We can add that in a follow up.

Bug: 107934
Fixed: 114413

Change-Id: I8d2d05edc435fedb45d490dd7686543fb0e1268a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756342
Reviewed-by: Simon Shields <simonshields@google.com>
Commit-Queue: Alex Legg <alexlegg@google.com>