[fidl][llcpp] Fix flexible inline enum encode bug
This bug is fairly well-described in the attached bug, and the fix is
straightforward: use the inline "flag" bit to check if an inline-able
flexible envelope has been set, rather than always casting to a ptr and
comparing to the always (on Fuchsia, anyway) 8-byte nullptr as was done
before. This is because the top four bytes of the inline envelope may
actually be set to a non-zero value as their "default", if those bytes
represent an inlined flexible enum.

Test: fx test conformance
Bug: 114125
Fixes: 114125
Change-Id: I9bfc9491a8f91069fb3daedacd8fbdedb0f1ebe5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756131
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Alex Zaslavsky <azaslavsky@google.com>