[fdf] Make magma GPU device non bindable
Add `DEVICE_ADD_NON_BINDABLE` to `magma_gpu`'s device flags. Currently,
no driver binds to it and making it non-bindable will make it operate
correctly in the DFv2 compat shim. This is because the DFv2 compat shim
does not allow adding child devices to a device that is not marked
non-bindable and the driver that creates `magma_gpu` tries to add a
child device to `magma_gpu`, called `gpu-performance-counters`.

Bug: 114093
Change-Id: Ic2e2948d9d2469d02c1d8e743c315a0d25aed6e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/755743
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>