[ffx][target] Process mdns info for emulator instances
This adds reading  the txt records on the mdns response from
Fuchsia devices. These txt records are added when Fuchsia is running
in an emulator which is using user mode networking. In this case, the
emulator is running on the local network interface, and mapping ports
on the guest (Fuchsia) to unused ports on the host.

The overridden configuration includes text records that identify
the host IP address and port mappings being used so that tools
like `ffx` can discover emulator instances via mdns and connect
via the mapped ports.

This information is used to create ffx targets and use the host
IP address and ssh port from the txt record.

Bug: 113136
Change-Id: I73e1642d7389440189413b6937461291ff14a8b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/752143
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>