[paver][pinecrest] avbab/liabr translation
Implement avbab/liabr metadata translation. The main differences
between them are:

1. avb ab uses little endian for crc storage. libabr uses big endian.
2. avb ab doesn't have the concept of recovery. When both slots are
   marked unbootable. The typical decision is to hang. libabr will
   attempt to boot recovery instead.

For 1, adjust crc endian order when reading/writing storage. For 2, when
libabr marks both slot unbootable, restores the tries remaining of the
current bootloader slot to 1, so that it can be booted next time.
At the same time, set a special flag in the reserved field to indicate
that both slot are in unbootable state. When reading from storage, if
the flag is set, mark both slot unbootable before presenting the
metadata to libabr. These will be done in libabr io callback, so it's
agnostic to libabr.

Bug: b/255567130

Change-Id: Ifec27a16b2e60d1cca674b54399d9bdea1c39f2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753045
Reviewed-by: David Pursell <dpursell@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Yecheng Zhao <zyecheng@google.com>