[settings] Cancel in-progress jobs on stream end
When a client initiates a hanging get then closes the connection on
their end, the watch job previously did not receive a cancelation
signal. This causes it to continue to wait for a response to its listen
request to the setting controller, which does not go away until the
setting value changes. This causes a leak of the watch job that can
build up async tasks until the service crashes: fxb/84201

This change cancels jobs when the source completes instead of just
completing them. The difference is that a cancelation signal is sent to
any currently active job. Only the watch job currently uses the
cancelation signal, so regular Set requests are not impacted.

Fixed: 84201
Test: fx test setui_service_tests, manually verified on device with
      local logging of # of port packets (fxr/756136)
Change-Id: I5d55828d25abff1a98a79b7e11b19c3ad186726f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756141
Reviewed-by: Paul Faria <paulfaria@google.com>
Commit-Queue: William Xiao <wxyz@google.com>