[recovery] Move ui files to recovery-ui lib
As part of adding a new ui for recovery-ota, the new ui expects to make
use of library code to help separate dependencies. This is a pre-step to
split new code from existing code.

This change had to change the interface for proxy-view-assistant
making the ConsoleViewAssistant an optional argument so the library
could be compiled with feature debug_console enabled.
Moved test mocking to only apply with cfg(test).

No other functional changes are included.

Test: Verified ui on nelson for all existing recovery-fdr and
recovery-ota images
Bug: b/257736402

Change-Id: Ia1373e7288fae7d0bad7456e761dee5ca25a2734
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/754408
Commit-Queue: Mike Comfoltey <comfoltey@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>
Reviewed-by: Kevin Twidle <kpt@google.com>