[driver_runtime] Set shutdown callback ctx to initial dispatcher
The OutgoingDirectory destructor will check that it is happening
on the same thread it was created on.

Since the driver Stop hook (and hence OutgoingDirectory destructor)
will run in the driver shutdown callback, we need to ensure
that when we do the callback, it looks like we are running on
the initial dispatcher. In actuality, all dispatchers have already
been shutdown.

Change-Id: I79e445069744f410b8dff777151352bfbbaea6ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756762
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>