[swd][assembly] Migrate most uses of Path{,Buf} to Utf8Path{,Buf}
Much of the packaging and assembly uses of path requires them to be
UTF-8. Previously we pushed down the checks deep in the libraries, but
this instead switches over to using camino::Utf8Path{,Buf} at the CLI
and serialization boundaries to make it clear that these values need
to be UTF-8.

Change-Id: Ia276b3478c992142786f6771b313d5cfb1dbbc25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/755938
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>