[gnsdk] Add gn sdk rules for OOT Structured Config
This change includes significant re-working of the fuchsia_component
template and prepare_package_inputs.py in addition to porting over
the main structured config templates.

All fuchsia_component templates now either take a manifest label
as an argument or internally generate a manifest label if a raw
manifest file is passed. This is to afford creation of a manifest
label that can be passed to both the fuchsia_component template
and the structured config rules.

prepare_package_inputs.py previously handled copying v1 and v2
manifests in different ways, but it has been adjusted to be
CF version agnostic. The primary reason is because separating out
manifest label creation makes it impossible to put some of the
previously available values in `component_contents`. A secondary
benefit is that it's nice to have it not have branches for different
CF versions.

Bug: chromium:1255292
Change-Id: Ieebf583c68e64800382a03891fb8612dfb832647
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732324
Reviewed-by: Sébastien Marchand <sebmarchand@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Bryant Chandler <bryantchandler@google.com>