[fshost] launch zbi ramdisk when provided
When provided with a ramdisk in the boot items, decompress it and launch
it as a ramdisk so it can be matched as normal.

In addition to integration tests, this change was tested on a local
device by loading a "netboot" image, which utilizes a zbi ramdisk, and
ensuring the system booted properly.

Fixed: 109302
Test: Manually, via zedboot
Change-Id: I6d9b13336dde93b376f5ce3118a457f7dda6c0e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/754903
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>