[fx][build] s/"ZBI test"/"boot test" as a category
"ZBI test" is already a bit of a misnomer as it includes QEMU kernel
tests. Soon, we'll further want to generalize this class to include
UEFI application tests, which too will be ZBI-less in the general case.
Accordingly, we reframe the general umbrella of "ZBI tests" as
"boot tests" to better encapsulate these three categories. The concept
of a "ZBI test" still remains; it just no longer gets top billing on the
marquee.

We also rename `phys_qemu_test()` to `qemu_kernel_test()` for clarity.

We update the `zbi_tests` GN metadata as `boot_tests`. Although this
metadata is no longer load-bearing, it is still used by local tooling.
An upcoming change will extend this tooling to support UEFI tests for
local convenience in the short term - and further changes after that
will seek to properly harmonize it with what automated testing expects
(and in the process maybe migrate off it altogether).

Bug: 113961
Change-Id: Idd296bbf697102fdbaa88f80a85b357b91d42b9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/755925
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>