[dwc2] use temporary variable to store transferred length
Discovered a race condition when working on adb file transfer where
ep->req_offset is changed when we release the lock before we get to
complete the request. Save the transferred length in a variable before
we release the lock so we don't lose this value.

Test: tested on Vim3

Change-Id: Ifad88e02199656bf76e0f63c9ff391e77e0e66cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756282
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Ruby Zhuang <rdzhuang@google.com>