[emu] Remove dependency on "ip" binary
The `ffx emu` tool has an unstated dependency on the linux program "ip",
which it uses for retrieving details about the host's network interfaces
when determining if a Tun/Tap network can be used. This breaks on some
infra hosts that don't have that program installed.

This CL removes the tool's dependency on the external "ip" program by
replacing its usage with the nix::ifaddr library instead. This makes
system calls directly instead of using the external program to do the
same. All functionality is equivalent in the new version, and tests have
been updated for the new format.

Bug: 113985
Change-Id: I6af1922ad677131068519e0b9285c3705215bc2c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/755962
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>