[emu] Remove dead code from tuntap.rs
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/755962 removed an
external dependency on the "ip" binary for checking network capabilities
of the host. However, after code review it was noticed that we left
behind one of the translation structs for that program; the compiler did
not detect the dead code.

This CL removes the struct, the comments surrounding it, and a build
dependency on serde that was no longer needed.

Bug: 113985
Change-Id: I1d9ecc2cffd4be663ad34f761ace3221d0b5d1a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757772
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Fuchsia-Auto-Submit: Daniel Dougherty <doughertyda@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>