[netstack3] Check exact addr when allocating port
The BoundSocketMap code that checks whether an address is available was
checking conflicts with potential shadower and shadow-ee addresses, but
was not checking that the actual address itself wasn't occupied. This
resulted in failures to bind both TCP and UDP sockets when the port
number was 0, even though not all ephemeral ports were in use. As a
consequence, binding a socket when other bound sockets were present
would occasionally fail, manifesting as syscall test flakiness.

This patch adds tests for UDP and TCP to check that can each allocate
the last ephemeral port as well as a fix for the underlying bug that
allows those tests to pass.

Bug: 114752, 114785
Change-Id: Ic6ceecec12b394f633fc715c9893f94783132333
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757806
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>