[vdl] Add deprecation notice and sleep to `fx vdl`
This CL updates the deprecation notice in the `fx vdl` script to
indicate that the tool is slated for removal. It also adds a sleep
command to pause the script for 5 seconds to allow the user time to read
the message before execution continues.

Bug: 113140
Change-Id: I2e0469263cd349188f9076f06882eba4f77c97ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757774
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>