[build] Change hermetic deps rules for clang-doc.
Previously we allowlisted the one clang-doc invocation (a test) to allow
it to pass hermetic deps. As we start to expand the uses of clang-doc,
this approach will not be scalable. Instead, allowlist the prebuilt
headers (clang-doc implicitly pulls these in as it parses the headers
being documented).

This also matches the behavior with libraries and other prebuilt
artifacts already present.

Adds proper inputs for clang_doc invocation for the files in the
build.

Also add the compile_commands.json file with a reference to the GN bug
for why this is required.

Bug: 114829

Change-Id: I4e677b802f75299a860b6fe5de96ba9f9421bf75
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757742
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Shai Barack <shayba@google.com>