[fxtest] Remove support for components v1 tests
There are no longer any v1 tests in tree that need to be launched with
run-test-component, so remove all references to it. Note that any
options related to v1 are left for now but cause fx test to fail with
an error message. Similarly, if fx test encounters a cmx test in
tests.json it will note that the test is unsupported rather than
crashing.

Bug: 109179
Testing: fx test fxtest
Manually verify specifying --realm produces an error message
Verify that updating tests.json to contain a cmx test and attempting to
run it results in an error.

Change-Id: If3c1c9d03e47f83d8a176b7e1ce18630ec510a5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753885
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>