[blobfs] Add sync to fragmentation test.
Fixes a flake wherein not all reserved extents would be released in time
for the large blob to be added at the end. However, this would normally
fail to return the final extent reservation in time, so adding a sync
actually changes the most common behaviour, but at least it is stable
now. Had to update the expected values for the last check to match.

Fixed: 114802
Change-Id: Ia85efa696343efe974a5f3f3aa3d506fa9edce1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757733
Reviewed-by: Chris Tam <godtamit@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>