[debugger] Evaluate expressions as blocks.
Allows normal inputs to the expression system to contain multiple
statements. Removes the requirement that the last statement in a C block
end in a semicolon.

From the UI this allows things like:
  print int i = 0; i * 9
Previously this would have worked, but would have required {} around the
whole thing. This capability will be used to greater benefit in the
pretty printing system.

Updates error messages to be a bit more clear in this context.

Change-Id: I6d7a30bbcbbe27a79cba9fb7d33f22da7472e880
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757906
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>