[driver_runtime] Make fdf::Arena implement fdf::Arena.
It heap allocates memory, but will be freed when the
fdf::Arena C++ object is freed.

For now this is the same implementation as the heap
allocation part of fidl::Arena.

RELNOTES: |fdf::Arena| now implements |fidl::AnyArena| and
can be used with |fidl::ToWire|.

Bug: 114822

Change-Id: I4739457cf655a0ee3fea4cbafc650cf298631368
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758605
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>