[spi][core] Unregister VMOs on client close
With the new SPI device topology, DdkOpen/DdkClose should be implemented
for the Banjo device rather than the root device as was done previously.
This allows VMOs to be unregistered when a client closes the device,
freeing the VMO IDs for use by subsequent clients.

Test: Ran selina_driver multiple times, no VMO register errors were
      reported after the first run
Change-Id: I827f97fff61fb3716439eb861389b6794d1b45c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758262
Commit-Queue: Braden Kell <bradenkell@google.com>
Reviewed-by: Ian Fisher <iafisher@google.com>